Move direct runner test into direct runner directory.
